requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0   
anyio==4.6.0
asgiref==3.8.1
attrs==24.2.0
audioread==3.0.1
av==12.3.0
backoff==2.2.1
bcrypt==4.2.0
build==1.2.2
cachetools==5.5.0        
certifi==2024.8.30       
cffi==1.17.1
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6    
chromadb==0.5.7
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1      
contourpy==1.3.0
ctranslate2==4.4.0       
cycler==0.12.1
dataclasses-json==0.6.7  
decorator==5.1.1
Deprecated==1.2.14       
dirtyjson==1.0.8
distro==1.9.0
durationpy==0.7
fastapi==0.115.0
faster-whisper==1.0.3
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
google-ai-generativelanguage==0.6.6
google-api-core==2.20.0
google-api-python-client==2.146.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
groq==0.11.0
grpcio==1.66.1
grpcio-status==1.62.3
gTTS==2.5.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
kubernetes==31.0.0
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.5
langchain-google-genai==2.0.0
langchain-groq==0.2.0
langchain-huggingface==0.1.0
langchain-text-splitters==0.3.0
langsmith==0.1.126
lazy_loader==0.4
librosa==0.10.2.post1
llama-index-core==0.11.12
llama-parse==0.5.6
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
noisereduce==3.0.2
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.2
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
posthog==3.6.6
proto-plus==1.24.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pygame==2.6.0
Pygments==2.18.0
pyparsing==3.1.4
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.1
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.1.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
SQLAlchemy==2.0.35
starlette==0.38.6
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.4.1
tqdm==4.66.5
transformers==4.44.2
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.30.6
watchfiles==0.24.0
websocket-client==1.8.0
websockets==13.1
wrapt==1.16.0
yarl==1.12.1
zipp==3.20.2
